requirements.txt
# Core
numpy>=1.26
pandas>=2.1
pyarrow>=15.0
openpyxl>=3.1
scikit-learn>=1.4
matplotlib>=3.8
tqdm>=4.66
PyYAML>=6.0
joblib>=1.3
psutil>=5.9

# NLP / DL
torch>=2.3
transformers>=4.41
datasets>=2.19
accelerate>=0.30
peft>=0.11
sentencepiece>=0.2.0
safetensors>=0.4

# Optional (graceful fallback if not installed):
#   bitsandbytes        -> 4-bit QLoRA. Skip on RTX 50-series until upstream supports sm_120.
#   onnxruntime         -> CPU-side quantized ensemble inference
#   pynvml              -> GPU profiling (auto-skipped if absent)
#   lime                -> sentence-level explanations
bitsandbytes>=0.43; sys_platform != "win32"
onnxruntime>=1.18
pynvml>=11.5
lime>=0.2.0.1

# Misc
gitpython>=3.1
